raftstore: extract much of coprocessor::properties
