+++ components/raftstore/src/coprocessor/mod.rs
@@ -3,3 +3,4 @@ pub use error::{Error, Result};
 pub mod config;
 pub mod error;
 pub mod metrics;
+pub mod properties;
